MAT-6114 run test button changes
