+++ src/components/testCaseLanding/qdm/CreateCodeCoverageNavTabs.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { CircularProgress, Box } from "@mui/material";
 import { Button, Tabs, Tab } from "@madie/madie-design-system/dist/react";
 import AddIcon from "@mui/icons-material/Add";
 import FileUploadIcon from "@mui/icons-material/FileUpload";
@@ -8,6 +7,7 @@ import { Measure, MeasureErrorType, TestCase } from "@madie/madie-models";
 import { TestCasesPassingDetailsProps } from "../common/interfaces";
 import { useFeatureFlags } from "@madie/madie-util";
 import { useQdmExecutionContext } from "../../routes/qdm/QdmExecutionContext";
+import RunTestButton from "../common/runTestsButton/RunTestsButton";
 
 export interface NavTabProps {
   activeTab: string;
@@ -16,7 +16,7 @@ export interface NavTabProps {
   canEdit: boolean;
   measure: Measure;
   createNewTestCase: (value: string) => void;
-  executeTestCases: (value: string) => void;
+  executeTestCases: () => void;
   onImportTestCases?: () => void;
   testCasePassFailStats: TestCasesPassingDetailsProps;
   coveragePercentage: number;
@@ -36,7 +36,7 @@ const defaultStyle = {
 };
 
 export default function CreateCodeCoverageNavTabs(props: NavTabProps) {
-  const { executionContextReady, executing } = useQdmExecutionContext();
+  const { executionContextReady } = useQdmExecutionContext();
 
   const {
     activeTab,
@@ -73,6 +73,15 @@ export default function CreateCodeCoverageNavTabs(props: NavTabProps) {
       </div>
     );
   };
+
+  const hasErrors =
+    measure?.cqlErrors ||
+    measure?.errors?.includes(
+      MeasureErrorType.MISMATCH_CQL_POPULATION_RETURN_TYPES
+    ) ||
+    _.isNil(measure?.groups) ||
+    measure?.groups.length === 0 ||
+    _.isEmpty(validTestCases);
 
   return (
     <div
@@ -140,41 +149,11 @@ export default function CreateCodeCoverageNavTabs(props: NavTabProps) {
             New Test Case
           </Button>
         </div>
-        <div style={{ margin: "0 6px 0 26px" }}>
-          <Box sx={{ position: "relative" }}>
-            <Button
-              variant="cyan"
-              disabled={
-                !!measure?.cqlErrors ||
-                measure?.errors?.includes(
-                  MeasureErrorType.MISMATCH_CQL_POPULATION_RETURN_TYPES
-                ) ||
-                _.isNil(measure?.groups) ||
-                measure?.groups.length === 0 ||
-                !executionContextReady ||
-                executing ||
-                _.isEmpty(validTestCases)
-              }
-              onClick={executeTestCases}
-              data-testid="execute-test-cases-button"
-            >
-              Run Test Cases
-            </Button>
-            {executing && (
-              <CircularProgress
-                size={24}
-                sx={{
-                  color: "#209FA6",
-                  position: "absolute",
-                  top: "50%",
-                  left: "50%",
-                  marginTop: "-5px",
-                  marginLeft: "-12px",
-                }}
-              />
-            )}
-          </Box>
-        </div>
+        <RunTestButton
+          hasErrors={hasErrors}
+          isExecutionContextReady={executionContextReady}
+          onRunTests={executeTestCases}
+        />
       </div>
     </div>
   );
